Revert "[SPARK-22308] Support alternative unit testing styles in external applications"
This reverts commit 592cfeab9caeff955d115a1ca5014ede7d402907.
