+++ core/src/test/scala/org/apache/spark/SharedSparkContext.scala
@@ -29,23 +29,10 @@ trait SharedSparkContext extends BeforeAndAfterAll with BeforeAndAfterEach { sel
 
   var conf = new SparkConf(false)
 
-  /**
-   * Initialize the [[SparkContext]].  Generally, this is just called from beforeAll; however, in
-   * test using styles other than FunSuite, there is often code that relies on the session between
-   * test group constructs and the actual tests, which may need this session.  It is purely a
-   * semantic difference, but semantically, it makes more sense to call 'initializeContext' between
-   * a 'describe' and an 'it' call than it does to call 'beforeAll'.
-   */
-  protected def initializeContext(): Unit = {
-    if (null == _sc) {
-      _sc = new SparkContext(
-        "local[4]", "test", conf.set("spark.hadoop.fs.file.impl", classOf[DebugFilesystem].getName))
-    }
-  }
-
   override def beforeAll() {
     super.beforeAll()
-    initializeContext()
+    _sc = new SparkContext(
+      "local[4]", "test", conf.set("spark.hadoop.fs.file.impl", classOf[DebugFilesystem].getName))
   }
 
   override def afterAll() {
